requirements.txt
flask
flask-session
cs50
python-slugify